Refactor: 불필요한 부분 삭제 및 변경
