+++ PlanJ/app/src/main/java/com/boostcamp/planj/ui/main/CategoryFragment.kt
@@ -1,25 +1,18 @@
 package com.boostcamp.planj.ui.main
 
 import android.os.Bundle
-import android.text.Layout.Directions
-import android.util.Log
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import android.widget.Toast
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.viewModels
 import androidx.lifecycle.Lifecycle
 import androidx.lifecycle.lifecycleScope
 import androidx.lifecycle.repeatOnLifecycle
-import androidx.navigation.fragment.FragmentNavigatorDestinationBuilder
 import androidx.navigation.fragment.findNavController
-import com.boostcamp.planj.R
 import com.boostcamp.planj.data.model.Category
 import com.boostcamp.planj.databinding.FragmentCategoryBinding
 
-import com.boostcamp.planj.ui.category.CategoryDialog
-import com.boostcamp.planj.ui.category.CategoryState
 import com.boostcamp.planj.ui.main.adapter.CategoryListAdapter
 import dagger.hilt.android.AndroidEntryPoint
 import kotlinx.coroutines.flow.collectLatest
@@ -43,7 +36,6 @@ class CategoryFragment : Fragment() {
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
-        binding.lifecycleOwner = this
         binding.view = this
         initAdapter()
         setObserver()
