+++ PlanJ/app/src/main/java/com/boostcamp/planj/ui/main/CategoryViewModel.kt
@@ -4,8 +4,6 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.boostcamp.planj.data.model.Category
 import com.boostcamp.planj.data.repository.MainRepository
-import com.boostcamp.planj.ui.category.CategoryState
-import dagger.hilt.android.AndroidEntryPoint
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.SharingStarted
