+++ PlanJ/app/src/main/res/layout/dialog_add_schedule.xml
@@ -60,6 +60,7 @@
                 android:background="@drawable/dialog_category_backgroud"
                 android:inputType="none"
                 android:padding="10dp"
+                android:textStyle="bold"
                 android:textAlignment="center"
                 android:textColor="@color/black" />
 
@@ -76,6 +77,7 @@
             app:hintEnabled="false"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent"
+            app:boxStrokeColor="@color/main3"
             app:layout_constraintTop_toBottomOf="@id/til_dialog_schedule_input">
 
             <com.google.android.material.textfield.TextInputEditText
@@ -83,6 +85,7 @@
                 android:layout_width="match_parent"
                 android:layout_height="match_parent"
                 android:hint="제목"
+                android:textColorHint="@color/main3"
                 android:imeOptions="actionDone"
                 android:inputType="text"
                 android:maxLength="25"
