+++ PlanJ/app/src/main/res/layout/fragment_category.xml
@@ -12,11 +12,10 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        android:background="@color/white"
-        >
+        android:background="@color/white">
 
         <androidx.appcompat.widget.Toolbar
-            android:id="@+id/tb_category_main"
+            android:id="@+id/toolbar_category"
             android:layout_width="0dp"
             android:layout_height="?attr/actionBarSize"
             app:layout_constraintEnd_toEndOf="parent"
@@ -46,7 +45,7 @@
             app:layout_constraintBottom_toBottomOf="parent"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toBottomOf="@id/tb_category_main">
+            app:layout_constraintTop_toBottomOf="@id/toolbar_category">
 
             <androidx.constraintlayout.widget.ConstraintLayout
                 android:layout_width="match_parent"
@@ -56,11 +55,12 @@
                 <LinearLayout
                     android:id="@+id/layout_category_add_category"
                     android:layout_width="0dp"
-                    android:layout_height="wrap_content"
+                    android:layout_height="60dp"
                     android:layout_marginStart="8dp"
                     android:layout_marginTop="16dp"
                     android:layout_marginEnd="8dp"
                     android:background="@drawable/round_r8_stroke_w1"
+                    android:gravity="center_vertical"
                     android:onClick="@{() -> view.addCategoryDialog(``)}"
                     android:orientation="horizontal"
                     android:padding="16dp"
@@ -77,7 +77,7 @@
                         android:layout_width="wrap_content"
                         android:layout_height="wrap_content"
                         android:layout_marginStart="8dp"
-                        android:text="새로운 카터고리 생성"
+                        android:text="새로운 카테고리 생성"
                         android:textColor="@color/black"
                         android:textSize="16sp"
                         android:textStyle="bold" />
