+++ PlanJ/app/src/main/res/layout/fragment_category_detail.xml
@@ -14,7 +14,7 @@
         android:layout_height="match_parent">
 
         <androidx.appcompat.widget.Toolbar
-            android:id="@+id/tb_category_detail_main"
+            android:id="@+id/toolbar_category_detail"
             android:layout_width="0dp"
             android:layout_height="?attr/actionBarSize"
             app:layout_constraintEnd_toEndOf="parent"
@@ -44,7 +44,7 @@
             android:layout_height="0dp"
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintTop_toBottomOf="@id/tb_category_detail_main"
+            app:layout_constraintTop_toBottomOf="@id/toolbar_category_detail"
             app:layout_constraintBottom_toBottomOf="parent"
             >
