+++ PlanJ/app/src/main/res/layout/item_list_category.xml
@@ -1,8 +1,9 @@
 <?xml version="1.0" encoding="utf-8"?>
-<layout xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:android="http://schemas.android.com/apk/res/android">
+<layout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto">
 
     <data>
+
         <variable
             name="category"
             type="com.boostcamp.planj.data.model.Category" />
@@ -10,37 +11,36 @@
 
     <LinearLayout
         android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:orientation="horizontal"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        android:padding="8dp"
+        android:layout_height="60dp"
         android:layout_margin="8dp"
-        app:setCategoryBackground="@{category}"
-        >
+        android:gravity="center_vertical"
+        android:orientation="horizontal"
+        android:padding="8dp"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:setCategoryBackground="@{category}">
 
 
         <TextView
             android:layout_width="0dp"
             android:layout_height="wrap_content"
-            android:text="@{category.categoryName}"
-            android:textColor="@color/black"
             android:layout_marginStart="8dp"
             android:layout_weight="9"
-            android:textStyle="bold"
-            android:textSize="16sp" />
+            android:text="@{category.categoryName}"
+            android:textColor="@color/black"
+            android:textSize="16sp"
+            android:textStyle="bold" />
 
         <TextView
             android:id="@+id/tv_category_pop_up_menu"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:text="&#8942;"
-            android:textSize="16sp"
-            android:textAlignment="center"
             android:paddingHorizontal="10dp"
             android:paddingVertical="5dp"
-            app:isPopUpMenuVisible="@{category}"
-            />
+            android:text="&#8942;"
+            android:textAlignment="center"
+            android:textSize="16sp"
+            app:isPopUpMenuVisible="@{category}" />
 
 
     </LinearLayout>
