+++ PlanJ/app/src/main/res/values/themes.xml
@@ -10,8 +10,5 @@
 
     <style name="dialog" parent="Theme.AppCompat.Dialog">
         <item name="android:windowBackground">@android:color/transparent</item>
-
-        <!-- status bar 노랑색으로 설정 -->
-<!--        <item name="android:statusBarColor">#FFFF00</item>-->
     </style>
 </resources>
